requirements.txt
# Production Agent System - Python Dependencies

# Core Framework
pydantic>=2.5.0
pydantic-ai>=0.0.14
python-dotenv>=1.0.0

# Marketing Dashboard
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.1.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Slack
slack-sdk>=3.26.0
slack-bolt>=1.18.0

# AI/LLM
anthropic>=0.7.8
openai>=1.6.1

# MCP
mcp>=0.9.0

# Async
aiohttp>=3.9.1

# Utilities
jinja2>=3.1.2
click>=8.1.7
python-json-logger>=2.0.7
structlog>=23.2.0

# Observability (optional)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-sqlalchemy>=0.42b0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0
hypothesis>=6.92.1
freezegun>=1.4.0
testcontainers>=3.7.1
locust>=2.20.0
bandit>=1.7.5
safety>=2.3.5

# Development
black>=23.12.1
ruff>=0.1.8
mypy>=1.7.1
pre-commit>=3.6.0
